+++ price-adapter/src/mappers/band_static_mapper.rs
@@ -31,6 +31,9 @@ impl BandStaticMapper {
             "huobipro" => Ok(include_str!(concat!(
                 env!("CARGO_MANIFEST_DIR"),
                 "/resources/huobipro.json"
+            "coinmarketcap" => Ok(include_str!(concat!(
+                env!("CARGO_MANIFEST_DIR"),
+                "/resources/coinmarketcap.json"
             ))),
             _ => Err(Error::UnsupportedSource), // Add more matches for other sources
         }?;