+++ price-adapter/src/sources.rs
@@ -2,8 +2,10 @@ mod band_stable_coin;
 mod binance;
 mod coingecko;
 mod huobipro;
+mod coinmarketcap;
 
 pub use band_stable_coin::*;
 pub use binance::*;
 pub use coingecko::*;
 pub use huobipro::*;
+pub use coinmarketcap::*;