+++ price-adapter-raw/CHANGELOG.md
@@ -5,3 +5,4 @@
 - Support CoinGecko sources
 - Support Binance Websocket source
 - Support Huobipro sources
+- Support CoinMarketCap source