+++ price-adapter/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Support CoinGecko sources
+- Support CoinMarketCap source
 - Support Binance Websocket source
 - Support HuobiPro sources
 - Create Interval and Websocket service to manage sources